Rename types package to helpers and add detector
