+++ .eslintignore
@@ -7,3 +7,5 @@ node_modules
 dist
 package.json
 !.storybook
+
+packages/browser-wallet-api-helpers/lib
