+++ README.md
@@ -4,7 +4,7 @@ This repository contains the implementation of the Concordium browser wallet, wh
 
 -   [browser-wallet](./packages/browser-wallet) is the actual extension installed in chrome.
 -   [browser-wallet-api](./packages/browser-wallet-api) is the API injected into web pages, allowing easy integration with the extension.
--   [browser-wallet-api-types](./packages/browser-wallet-api-types) is types associated with the browser wallet api, which can be installed to add relevant types to the project. This is published as a package on NPM as [@concordium/browser-wallet-api-types](https://www.npmjs.com/package/@concordium/browser-wallet-api-types).
+-   [browser-wallet-api-helpers](./packages/browser-wallet-api-helpers) is types and helper methods associated with the browser wallet api, which can be installed to projects that want to utilize the browser wallet API. This is published as a package on NPM as [@concordium/browser-wallet-api-helpers](https://www.npmjs.com/package/@concordium/browser-wallet-api-helpers).
 -   [browser-wallet-message-hub](./packages/browser-wallet-message-hub) is utilities used for internal communication between the separate contexts in the chrome extension.
 
 ## Installing
