+++ examples/piggybank/package.json
@@ -2,7 +2,7 @@
     "name": "piggybank",
     "packageManager": "yarn@3.2.0",
     "dependencies": {
-        "@concordium/browser-wallet-api-types": "workspace:^",
+        "@concordium/browser-wallet-api-helpers": "workspace:^",
         "@concordium/web-sdk": "^0.1.2",
         "react": "^18.1.0",
         "react-dom": "^18.1.0"
