+++ examples/piggybank/src/Root.tsx
@@ -12,7 +12,7 @@ import {
     UpdateContractPayload,
 } from '@concordium/web-sdk';
 
-import { WalletApi } from '@concordium/browser-wallet-api-types';
+import { detectConcordiumProvider } from '@concordium/browser-wallet-api-helpers';
 import PiggyIcon from './assets/piggy-bank-solid.svg';
 import HammerIcon from './assets/hammer-solid.svg';
 
@@ -43,39 +43,6 @@ const isPiggybankSmashed = (state: PiggyBankState): state is PiggyBankStateSmash
     (state as PiggyBankStateSmashed).Smashed !== undefined;
 
 const client = new JsonRpcClient(new HttpProvider(JSON_RPC_URL));
-
-/**
- * Promise resolves to the Concordium provider when it has been successfully injected into
- * the window and is ready for use.
- * @param timeout determines how long to wait before rejecting if the Concordium provider is not available, in milliseconds.
- * @returns a promise containing the Concordium Wallet provider API.
- */
-// TODO This function should be made available from the web-sdk for ease of use.
-async function detectConcordiumProvider(timeout = 5000): Promise<WalletApi> {
-    return new Promise((resolve, reject) => {
-        if (window.concordium) {
-            resolve(window.concordium);
-        } else {
-            const t = setTimeout(() => {
-                if (window.concordium) {
-                    resolve(window.concordium);
-                } else {
-                    reject();
-                }
-            }, timeout);
-            window.addEventListener(
-                'concordium#initialized',
-                () => {
-                    if (window.concordium) {
-                        clearTimeout(t);
-                        resolve(window.concordium);
-                    }
-                },
-                { once: true }
-            );
-        }
-    });
-}
 
 /**
  * Global application state.
