+++ packages/browser-wallet-api-helpers/CHANGELOG.md
@@ -1,0 +1,6 @@
+# Changelog
+
+## 0.1.0
+
+-   Initialized from the old browser-wallet-api-types package.
+-   Added method for detecting the injected Concordium browser wallet API.
