+++ packages/browser-wallet-api-helpers/src/index.ts
@@ -1,0 +1,2 @@
+export * from './wallet-api-types';
+export { detectConcordiumProvider } from './detector';
