+++ packages/browser-wallet-api-helpers/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../../tsconfig.json",
+    "compilerOptions": {
+        "declaration": true,
+        "outDir": "./lib"
+    },
+    "include": ["./src"]
+}
