--- packages/browser-wallet-api-types/CHANGELOG.md
@@ -1,9 +1,0 @@
-# Changelog
-
-## 0.1.1
-
--   Updated README
-
-## 0.1.0
-
--   Initial release
