--- packages/browser-wallet-api-types/tsconfig.json
@@ -1,4 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "include": ["**/*.d.ts"]
-}
