+++ packages/browser-wallet-api/package.json
@@ -6,7 +6,7 @@
     "author": "Concordium",
     "license": "Apache",
     "dependencies": {
-        "@concordium/browser-wallet-api-types": "workspace:^",
+        "@concordium/browser-wallet-api-helpers": "workspace:^",
         "@concordium/web-sdk": "^0.1.2"
     },
     "devDependencies": {
