+++ packages/browser-wallet-api/tsconfig.json
@@ -3,7 +3,8 @@
     "compilerOptions": {
         "baseUrl": ".",
         "paths": {
-            "@concordium/browser-wallet-message-hub": ["../../packages/browser-wallet-message-hub/src"]
+            "@concordium/browser-wallet-message-hub": ["../../packages/browser-wallet-message-hub/src"],
+            "@concordium/browser-wallet-api-helpers": ["../browser-wallet-api-helpers/src"]
         },
         "outDir": "./lib"
     },
