+++ packages/browser-wallet/package.json
@@ -13,7 +13,7 @@
     },
     "dependencies": {
         "@concordium/browser-wallet-api": "workspace:^",
-        "@concordium/browser-wallet-api-types": "workspace:^",
+        "@concordium/browser-wallet-api-helpers": "workspace:^",
         "@concordium/web-sdk": "^0.1.2",
         "buffer": "^6.0.3",
         "check-password-strength": "^2.0.5",
